+++ packages/js-toolkit/Base/managers/RefsManager.ts
@@ -7,6 +7,7 @@ import {
   endsWith,
   startsWith,
 } from '../../utils/index.js';
+import { features } from '../features.js';
 
 const NORMALIZE_REF_NAME_REGEX = /\[\]$/;
 const NAMESPACED_REF_REGEX = /^(.*\.)/;
@@ -38,12 +39,15 @@ export function normalizeRefName(name: string) {
  * Filter refs belonging to the related Base instance.
  */
 function refBelongToInstance(ref: HTMLElement, rootElement: HTMLElement, name: string) {
-  const isPrefixed = startsWith(ref.dataset.ref, name);
+  const attributes = features.get('attributes');
+  const isPrefixed = startsWith(ref.getAttribute(attributes.ref), name);
   const firstComponentAncestor = getAncestorWhereUntil(
     ref,
     (el) =>
       el &&
-      (isPrefixed ? el.dataset.component === name && el : isDefined(el.dataset.component)) &&
+      (isPrefixed
+        ? el.getAttribute(attributes.component) === name && el
+        : el.getAttribute(attributes.component)) &&
       el !== rootElement,
     (el) => el === rootElement,
   );
@@ -62,36 +66,26 @@ function __register(that: RefsManager, refName: string) {
   const isMultiple = endsWith(refName, '[]');
   const propName = normalizeRefName(refName);
   const { name } = that.__base.$options;
+  const attributes = features.get('attributes');
 
   const refs = Array.from(
     that.__element.querySelectorAll<HTMLElement>(
-      `[data-ref="${refName}"],[data-ref="${name}.${refName}"]`,
+      `[${attributes.ref}="${refName}"],[${attributes.ref}="${name}.${refName}"]`,
     ),
   ).filter((ref) => refBelongToInstance(ref, that.__element, name));
 
+  /* v8 ignore next 7 */
   if (isDev && !isMultiple && refs.length > 1) {
     console.warn(
-      // @ts-ignore
-      `[${that.__base.$options.name}]`,
+      `[${that.__base.$id}]`,
       `The "${refName}" ref has been found multiple times.`,
       'Did you forgot to add the `[]` suffix to its name?',
     );
   }
 
-  if (!isMultiple && refs.length <= 1 && !isDefined(refs[0])) {
-    if (isDev) {
-      console.warn(
-        // @ts-ignore
-        `[${that.__base.$options.name}]`,
-        `The "${refName}" ref is missing.`,
-        `Is there an \`[data-ref="${refName}"]\` element in the component's scope?`,
-      );
-    }
-
-    return;
+  if (refs.length) {
+    that.__eventsManager.bindRef(refName, refs);
   }
-
-  that.__eventsManager.bindRef(refName, refs);
 
   Object.defineProperty(that, propName, {
     // @todo remove usage of non-multiple refs as array