+++ packages/tests/Base/managers/ChildrenManager.spec.ts
@@ -1,15 +1,10 @@
-import { describe, it, expect, mock, beforeEach, afterEach } from 'bun:test';
+import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { Base, BaseConfig } from '@studiometa/js-toolkit';
 import { getComponentElements } from '#private/Base/utils.js';
-import { h, useFakeTimers, useRealTimers, advanceTimersByTimeAsync } from '#test-utils';
+import { h } from '#test-utils';
 
 beforeEach(() => {
   document.body.innerHTML = '';
-  useFakeTimers();
-});
-
-afterEach(() => {
-  useRealTimers();
 });
 
 describe('The component resolution', () => {
@@ -50,7 +45,7 @@ describe('The component resolution', () => {
     const async = h('div', { dataComponent: 'AsyncComponent' });
     const div = h('div', {}, [async]);
 
-    const fn = mock();
+    const fn = vi.fn();
     class AsyncComponent extends Base {
       static config = {
         name: 'AsyncComponent',
@@ -72,8 +67,7 @@ describe('The component resolution', () => {
     }
 
     const component = new Component(div);
-    component.$mount();
-    await advanceTimersByTimeAsync(1);
+    await component.$mount();
     const asyncInstances = await Promise.all(component.$children.AsyncComponent);
     expect(asyncInstances[0]).toBeInstanceOf(Base);
     expect(fn).toHaveBeenCalledTimes(1);