+++ packages/tests/services/index.spec.ts
@@ -1,18 +1,17 @@
-import { test, expect } from 'bun:test';
-import * as toolkit from '@studiometa/js-toolkit';
-import { getFilenamesInFolder } from '#test-utils';
-
-const services = Object.fromEntries(
-  Object.entries(toolkit).filter(([name]) => name.startsWith('use')),
-);
-
-function capitalize(str) {
-  return str.replace(/^\w/, (c) => c.toUpperCase());
-}
+import { test, expect } from 'vitest';
+import * as services from '#private/services';
 
 test('components exports', () => {
-  const names = getFilenamesInFolder('../../js-toolkit/services/', import.meta.url)
-    .filter((name) => name !== 'Service')
-    .map((name) => `use${capitalize(name)}`);
-  expect(Object.keys(services).toSorted()).toEqual(names.toSorted());
+  expect(Object.keys(services).toSorted()).toMatchInlineSnapshot(`
+    [
+      "useDrag",
+      "useKey",
+      "useLoad",
+      "usePointer",
+      "useRaf",
+      "useResize",
+      "useScroll",
+      "useService",
+    ]
+  `);
 });