+++ packages/tests/utils/collide/collideCircleRect.spec.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from 'bun:test';
+import { describe, it, expect } from 'vitest';
 import { collideCircleRect } from '@studiometa/js-toolkit/utils';
 
 describe('collideCircleRect method', () => {